pyproject.toml
[tool.poetry]
name = "jp-nlp-tools"
version = "0.1.0"
description = ""
license = "MIT"
authors = ["Katsuya Shimabukuro <katsu.generation.888@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
mecab-python3 = "^0.996.5"
scikit-learn = "^0.23.1"
numpy = "^1.18.5"

[tool.poetry.dev-dependencies]
nose = "^1.3.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
